State variables
-Variable Location:
storage: Variables stored in storage are permanently stored on the blockchain and their state is preserved between function calls. They cannot be passed or returned to/from functions by default.

memory: Variables stored in memory are temporary and only exist during the execution of a function. Their state is not preserved after the function execution is complete. They can be passed and returned to/from functions.

calldata: Variables stored in calldata are read-only and hold data passed to the function from external calls. They are used for function arguments and cannot be modified.

-State variables are stored in smart contracts,

-Functions 2types
1.Creates transaction
2.Does not creat transaction(view or pure)

-Sol creates getters for public state variables

-IPFS+ smart contracts to store data